airplane/requirements.txt
# For airplane, the lowest-level directory that has requirements.txt in it defines the root directory.
# If you want access to pyshared, do NOT put requirements.txt in your task's directory.
# Instead, stick your requirements here in this common requirements.txt file.

# Required for returning any value from Python main function with local airplane run
airplanesdk
boto3>=1.20.20
coolname>=1.1.0
GitPython
notional==0.4.0
python-dateutil
pytz
pyyaml>=6.0
tenacity>=8.0.1
pendulum
ruamel.yaml>=0.17.20
ruamel.yaml.clib>=0.2.6
semver
